requirements.txt
aiogram==3.10.0
python-dotenv==1.0.1
aiohttp==3.9.5
xxtea
